requirements.txt
#GCP
google-cloud-storage
google-cloud-bigquery

#dataviz
matplotlib
seaborn

pygeohash
pytest

db-dtypes
numpy==1.23.5
pandas==1.5.3
scipy==1.10.0
scikit-learn==1.2.1

python-dateutil
